Refactor: fleshout types in commons crate
