+++ src/client/src/events/mod.rs
@@ -8,7 +8,7 @@ use tracer_aws::aws_metadata::get_aws_instance_metadata;
 use tracer_aws::config::PricingClient;
 use tracer_aws::types::pricing::EC2FilterBuilder;
 use tracer_common::debug_log::Logger;
-use tracer_common::event::attributes::system_metrics::SystemProperties;
+use tracer_common::types::event::attributes::system_metrics::SystemProperties;
 use tracer_extracts::metrics::SystemMetricsCollector;
 use tracing::info;
