+++ src/common/src/lib.rs
@@ -1,5 +1,3 @@
-pub mod event;
-pub mod pipeline_tags;
 pub mod recorder; // todo: this is very ugly, please move me out to tracer client
 
 pub mod constants;
@@ -7,3 +5,4 @@ pub mod debug_log;
 pub mod http_client;
 pub mod target_process;
 pub mod types;
+pub mod utils;
