+++ src/common/src/types/extracts/mod.rs
@@ -1,0 +1,1 @@
+pub mod db;
