+++ src/common/src/types/mod.rs
@@ -1,3 +1,7 @@
+pub mod event;
+pub mod extracts;
+pub mod pipeline_tags;
+
 use std::sync::Arc;
 use tokio::sync::RwLock;
