+++ src/extracts/src/syslog/mod.rs
@@ -11,10 +11,10 @@ use sysinfo::System;
 use tokio::sync::RwLock;
 use tokio_stream::StreamExt;
 use tracer_common::debug_log::Logger;
-use tracer_common::event::attributes::syslog::SyslogProperties;
-use tracer_common::event::attributes::EventAttributes;
-use tracer_common::event::ProcessStatus;
 use tracer_common::recorder::EventRecorder;
+use tracer_common::types::event::attributes::syslog::SyslogProperties;
+use tracer_common::types::event::attributes::EventAttributes;
+use tracer_common::types::event::ProcessStatus;
 
 const LINES_BEFORE: usize = 2;
